renamed controller implemented chargebee endpoints
